Separate WarHUD from Towny HUDManager.
